[BOTF-290] Sticky headers not working (#280)
## [BOTF-290] Description
Changed back to the sticky position approach and also Created the account menu for mobile screens
