+++ src/app/dashboard/login/login.page.scss
@@ -7,5 +7,6 @@
    .page-content {
       display: flex;
       align-items: center;
+      min-height: inherit;
    }
 }
