+++ src/components/common/texturePaperEllipses/TexturePaperEllipses.scss
@@ -4,6 +4,7 @@
 
    .content-wrap {
       position: relative;
+      width: 100%;
    }
 
    @keyframes rot1 {
