+++ src/components/content/home/Home.scss
@@ -1,12 +1,3 @@
-.home-page {
-   .page-content {
-      @media screen and (min-width: $screen-m) {
-         $header-height: 5.32rem;
-         padding-top: $header-height;
-      }
-   }
-}
-
 .home-content {
    .content-section {
       .section-header {
